Extract AMP boilerplate component to own file
