+++ assets/src/edit-story/output/story.js
@@ -24,38 +24,8 @@ import PropTypes from 'prop-types';
  */
 import StoryPropTypes from '../types';
 import getUsedAmpExtensions from './getUsedAmpExtensions';
-import { OutputPage } from './index';
-
-/**
- * Renders AMP boilerplate
- *
- * @see https://amp.dev/documentation/guides-and-tutorials/learn/spec/amp-boilerplate/
- * @see https://amp.dev/documentation/components/amp-story/#boilerplate
- *
- * @return {Element} AMP boilerplate.
- */
-function Boilerplate() {
-  return (
-    <>
-      <style
-        amp-boilerplate=""
-        dangerouslySetInnerHTML={{
-          __html:
-            'body{-webkit-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-moz-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-ms-animation:-amp-start 8s steps(1,end) 0s 1 normal both;animation:-amp-start 8s steps(1,end) 0s 1 normal both}@-webkit-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-moz-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-ms-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-o-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}',
-        }}
-      />
-      <noscript>
-        <style
-          amp-boilerplate=""
-          dangerouslySetInnerHTML={{
-            __html:
-              'body{-webkit-animation:none;-moz-animation:none;-ms-animation:none;animation:none}',
-          }}
-        />
-      </noscript>
-    </>
-  );
-}
+import Boilerplate from './ampBoilerplate';
+import { OutputPage } from './';
 
 function CustomCSS() {
   return (
